Clean up the TempCheck diagnostic example
I originally wrote this up to show the difference between instance and
class checks. Then, another person expanded upon this check. Those were
good canges, but the underlying example wasn't clearly fleshed out. This
should just make it so that the example is better understood. I think it
was confusing sending in a `temperature` parameter when it is really
just a temperature "limit". This should at least make it a little easier
to understand.